// src/movies/interfaces/swapi-response.interface.ts
export interface ApiMovie {
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  opening: string;
  budget: number;
  revenue: number;
  runtime: number;
  name: string;
  created: string;
  edited: string;
  id: number;
}